+++ spec/nim/tdebug_switch_user.nim
@@ -6,5 +6,5 @@ import auxiliary/test_utils
 
 let r = DebugSwitchUser.fromFile("src/nav_parent_switch.bin")
 
-assert r.code == uint8(1)
-assert (DebugSwitchUser_One(r.data)).val == int16(-190)
+assert r.code == 1
+assert (DebugSwitchUser_One(r.data)).val == -190
