+++ spec/nim/tprocess_xor_const.nim
@@ -6,5 +6,5 @@ import auxiliary/test_utils
 
 let r = ProcessXorConst.fromFile("src/process_xor_1.bin")
 
-assert r.key == uint8(255)
+assert r.key == 255
 assert r.buf == @[102'u8, 111'u8, 111'u8, 32'u8, 98'u8, 97'u8, 114'u8]
