+++ spec/nim/tswitch_else_only.nim
@@ -6,6 +6,6 @@ import auxiliary/test_utils
 
 let r = SwitchElseOnly.fromFile("src/switch_opcodes.bin")
 
-assert r.opcode == int8(83)
-assert r.primByte == int8(102)
+assert r.opcode == 83
+assert r.primByte == 102
 assert r.ut.value == @[114'u8, 0'u8, 73'u8, 66'u8]
